+++ Examples/ios_example/App/Test/MMFileScanTest.mm
@@ -10,7 +10,6 @@
 @implementation MMFileScanTest
 
 - (void)test {
-    
     NSUserDefaults *userDefaults = [NSUserDefaults standardUserDefaults];
     NSLog(@"%@", userDefaults);
 }