+++ Examples/ios_example/App/Test/cpp_file_scan_test.hpp
@@ -1,0 +1,16 @@
+//
+//  cpp_file_scan_test.hpp
+//  App
+//
+//  Created by crasowas on 2024/4/23.
+//
+
+#ifndef cpp_file_scan_test_hpp
+#define cpp_file_scan_test_hpp
+
+#include <stdio.h>
+#include <sys/stat.h>
+
+void test(const char *path);
+
+#endif /* cpp_file_scan_test_hpp */